# Avoid UDP cipher search if a NAT mapping exists