--- a/shadowsocks/cipher_list_test.go
+++ b/shadowsocks/cipher_list_test.go
@@ -17,6 +17,8 @@ package shadowsocks
 import (
 	"container/list"
 	"crypto/cipher"
+	"math/rand"
+	"net"
 	"testing"
 
 	"github.com/shadowsocks/go-shadowsocks2/shadowaead"
@@ -86,4 +88,45 @@ func TestCompatibleCiphers(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
+}
+
+func BenchmarkLocking(b *testing.B) {
+	var ip net.IP
+
+	ciphers, _ := MakeTestCiphers(MakeTestSecrets(1))
+	b.ResetTimer()
+	b.RunParallel(func(pb *testing.PB) {
+		for pb.Next() {
+			_, entries := ciphers.SnapshotForClientIP(nil)
+			ciphers.MarkUsedByClientIP(entries[0], ip)
+		}
+	})
+}
+
+func BenchmarkSnapshot(b *testing.B) {
+	// Create a list of cipher entries in a random order.
+
+	// Small cipher lists (N~1e3) fit entirely in cache, and are ~10 times
+	// faster to copy (per entry) than very large cipher lists (N~1e5).
+	const N = 1e3
+	ciphers, _ := MakeTestCiphers(MakeTestSecrets(N))
+
+	// Shuffling simulates the behavior of a real server, where successive
+	// ciphers are not expected to be nearby in memory.
+	_, entries := ciphers.SnapshotForClientIP(nil)
+	rand.Shuffle(N, func(i, j int) {
+		entries[i], entries[j] = entries[j], entries[i]
+	})
+	for _, entry := range entries {
+		// Reorder the list to match the shuffle
+		// (actually in reverse, but it doesn't matter).
+		ciphers.MarkUsedByClientIP(entry, nil)
+	}
+
+	b.ResetTimer()
+	b.RunParallel(func(pb *testing.PB) {
+		for pb.Next() {
+			ciphers.SnapshotForClientIP(nil)
+		}
+	})
 }