--- a/shadowsocks/udp.go
+++ b/shadowsocks/udp.go
@@ -163,16 +163,43 @@ func (s *udpService) Serve(clientConn net.PacketConn) error {
 				defer logger.Debugf("UDP(%v): done", clientAddr)
 				logger.Debugf("UDP(%v): Outbound packet has %d bytes", clientAddr, clientProxyBytes)
 			}
-			unpackStart := time.Now()
-			ip := clientAddr.(*net.UDPAddr).IP
-			buf, keyID, cipher, err := unpack(ip, textBuf, cipherBuf[:clientProxyBytes], s.ciphers)
-			timeToCipher = time.Now().Sub(unpackStart)
 
-			if err != nil {
-				return onet.NewConnectionError("ERR_CIPHER", "Failed to upack data from client", err)
+			cipherData := cipherBuf[:clientProxyBytes]
+			var textData []byte
+			targetConn := nm.Get(clientAddr.String())
+			if targetConn == nil {
+				clientLocation, locErr := s.m.GetLocation(clientAddr)
+				if locErr != nil {
+					logger.Warningf("Failed location lookup: %v", locErr)
+				}
+				debugUDPAddr(clientAddr, "Got location \"%s\"", clientLocation)
+
+				ip := clientAddr.(*net.UDPAddr).IP
+				var cipher shadowaead.Cipher
+				unpackStart := time.Now()
+				textData, keyID, cipher, err = unpack(ip, textBuf, cipherData, s.ciphers)
+				timeToCipher = time.Now().Sub(unpackStart)
+
+				if err != nil {
+					return onet.NewConnectionError("ERR_CIPHER", "Failed to unpack initial packet", err)
+				}
+
+				udpConn, err := net.ListenPacket("udp", "")
+				if err != nil {
+					return onet.NewConnectionError("ERR_CREATE_SOCKET", "Failed to create UDP socket", err)
+				}
+				targetConn = nm.Add(clientAddr, clientConn, cipher, udpConn, clientLocation, keyID)
+			} else {
+				unpackStart := time.Now()
+				textData, err = shadowaead.Unpack(textBuf, cipherData, targetConn.cipher)
+				timeToCipher = time.Now().Sub(unpackStart)
+				if err != nil {
+					return onet.NewConnectionError("ERR_CIPHER", "Failed to unpack data from client", err)
+				}
 			}
+			clientLocation = targetConn.clientLocation
 
-			tgtAddr := socks.SplitAddr(buf)
+			tgtAddr := socks.SplitAddr(textData)
 			if tgtAddr == nil {
 				return onet.NewConnectionError("ERR_READ_ADDRESS", "Failed to get target address", nil)
 			}
@@ -185,25 +212,8 @@ func (s *udpService) Serve(clientConn net.PacketConn) error {
 				return err
 			}
 
-			payload := buf[len(tgtAddr):]
-
-			targetConn := nm.Get(clientAddr.String())
-			if targetConn == nil {
-				clientLocation, locErr := s.m.GetLocation(clientAddr)
-				if locErr != nil {
-					logger.Warningf("Failed location lookup: %v", locErr)
-				}
-				debugUDPAddr(clientAddr, "Got location \"%s\"", clientLocation)
-
-				udpConn, err := net.ListenPacket("udp", "")
-				if err != nil {
-					return onet.NewConnectionError("ERR_CREATE_SOCKET", "Failed to create UDP socket", err)
-				}
-				targetConn = nm.Add(clientAddr, clientConn, cipher, udpConn, clientLocation, keyID)
-			}
-			clientLocation = targetConn.clientLocation
-
 			debugUDPAddr(clientAddr, "Proxy exit %v", targetConn.LocalAddr())
+			payload := textData[len(tgtAddr):]
 			proxyTargetBytes, err = targetConn.WriteTo(payload, tgtUDPAddr) // accept only UDPAddr despite the signature
 			if err != nil {
 				return onet.NewConnectionError("ERR_WRITE", "Failed to write to target", err)
@@ -237,6 +247,7 @@ func isDNS(addr net.Addr) bool {
 
 type natconn struct {
 	net.PacketConn
+	cipher shadowaead.Cipher
 	// We store the client location in the NAT map to avoid recomputing it
 	// for every downstream packet in a UDP-based connection.
 	clientLocation string
@@ -317,9 +328,10 @@ func (m *natmap) Get(key string) *natconn {
 	return m.keyConn[key]
 }
 
-func (m *natmap) set(key string, pc net.PacketConn, clientLocation string) *natconn {
+func (m *natmap) set(key string, pc net.PacketConn, cipher shadowaead.Cipher, clientLocation string) *natconn {
 	entry := &natconn{
 		PacketConn:     pc,
+		cipher:         cipher,
 		clientLocation: clientLocation,
 		defaultTimeout: m.timeout,
 	}
@@ -344,12 +356,12 @@ func (m *natmap) del(key string) net.PacketConn {
 }
 
 func (m *natmap) Add(clientAddr net.Addr, clientConn net.PacketConn, cipher shadowaead.Cipher, targetConn net.PacketConn, clientLocation, keyID string) *natconn {
-	entry := m.set(clientAddr.String(), targetConn, clientLocation)
+	entry := m.set(clientAddr.String(), targetConn, cipher, clientLocation)
 
 	m.metrics.AddUDPNatEntry()
 	m.running.Add(1)
 	go func() {
-		timedCopy(clientAddr, clientConn, cipher, entry, keyID, m.metrics)
+		timedCopy(clientAddr, clientConn, entry, keyID, m.metrics)
 		m.metrics.RemoveUDPNatEntry()
 		if pc := m.del(clientAddr.String()); pc != nil {
 			pc.Close()
@@ -378,14 +390,14 @@ func (m *natmap) Close() error {
 var maxAddrLen int = len(socks.ParseAddr("[2001:db8::1]:12345"))
 
 // copy from target to client until read timeout
-func timedCopy(clientAddr net.Addr, clientConn net.PacketConn, cipher shadowaead.Cipher, targetConn *natconn,
+func timedCopy(clientAddr net.Addr, clientConn net.PacketConn, targetConn *natconn,
 	keyID string, sm metrics.ShadowsocksMetrics) {
 	// pkt is used for in-place encryption of downstream UDP packets, with the layout
 	// [padding?][salt][address][body][tag][extra]
 	// Padding is only used if the address is IPv4.
 	pkt := make([]byte, udpBufSize)
 
-	saltSize := cipher.SaltSize()
+	saltSize := targetConn.cipher.SaltSize()
 	// Leave enough room at the beginning of the packet for a max-length header (i.e. IPv6).
 	bodyStart := saltSize + maxAddrLen
 
@@ -429,7 +441,7 @@ func timedCopy(clientAddr net.Addr, clientConn net.PacketConn, cipher shadowaead
 			//           [            packBuf             ]
 			//           [          buf           ]
 			packBuf := pkt[saltStart:]
-			buf, err := shadowaead.Pack(packBuf, plaintextBuf, cipher) // Encrypt in-place
+			buf, err := shadowaead.Pack(packBuf, plaintextBuf, targetConn.cipher) // Encrypt in-place
 			if err != nil {
 				return onet.NewConnectionError("ERR_PACK", "Failed to pack data to client", err)
 			}