@bemasc 2020-09-12
@JonathanDCohen FYI, this corresponds to some discussions we've had.